chatbot_test/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
colorama==0.4.6
et-xmlfile==1.1.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
numpy==1.26.1
openai==0.28.1
openpyxl==3.1.2
pandas==2.1.2
pandas-stubs==2.1.1.230928
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tqdm==4.66.1
types-pytz==2023.3.1.1
tzdata==2023.3
urllib3==2.0.7
yarl==1.9.2
